+++ backend/canisters/group_index/impl/src/lib.rs
@@ -136,7 +136,14 @@ struct Data {
     pub cached_metrics: CachedMetrics,
     pub local_index_map: LocalGroupIndexMap,
     pub fire_and_forget_handler: FireAndForgetHandler,
+    // TODO: Remove serde default
+    #[serde(default = "video_call_operators")]
+    pub video_call_operators: Vec<Principal>,
     pub rng_seed: [u8; 32],
+}
+
+fn video_call_operators() -> Vec<Principal> {
+    vec![Principal::from_text("nmufs-fiu7o-cyg5v-ozcjx-b5qsb-y6nsy-viid6-esfxk-s4nzb-yv2u3-jae").unwrap()]
 }
 
 impl Data {
@@ -150,6 +157,7 @@ impl Data {
         cycles_dispenser_canister_id: CanisterId,
         proposals_bot_user_id: UserId,
         escrow_canister_id: CanisterId,
+        video_call_operators: Vec<Principal>,
         test_mode: bool,
     ) -> Data {
         Data {
@@ -176,6 +184,7 @@ impl Data {
             cached_metrics: CachedMetrics::default(),
             local_index_map: LocalGroupIndexMap::default(),
             fire_and_forget_handler: FireAndForgetHandler::default(),
+            video_call_operators,
             rng_seed: [0; 32],
         }
     }
@@ -280,6 +289,7 @@ impl Default for Data {
             cached_metrics: CachedMetrics::default(),
             local_index_map: LocalGroupIndexMap::default(),
             fire_and_forget_handler: FireAndForgetHandler::default(),
+            video_call_operators: Vec::default(),
             rng_seed: [0; 32],
         }
     }