+++ backend/canisters/user/api/src/updates/mod.rs
@@ -36,6 +36,8 @@ pub mod delete_direct_chat;
 pub mod delete_group;
 pub mod delete_messages;
 pub mod edit_message_v2;
+pub mod end_video_call;
+pub mod join_video_call;
 pub mod leave_community;
 pub mod leave_group;
 pub mod manage_favourite_chats;