+++ backend/canisters/user/impl/src/guards.rs
@@ -32,6 +32,14 @@ pub fn caller_is_owner_or_local_user_index() -> Result<(), String> {
     }
 }
 
+pub fn caller_is_owner_or_video_call_operator() -> Result<(), String> {
+    if read_state(|state| state.is_caller_owner() || state.is_caller_video_call_operator()) {
+        Ok(())
+    } else {
+        Err("Caller is not the canister owner or the video call operator".to_owned())
+    }
+}
+
 pub fn caller_is_group_index() -> Result<(), String> {
     if read_state(|state| state.is_caller_group_index()) {
         Ok(())
@@ -53,5 +61,13 @@ pub fn caller_is_known_group_or_community_canister() -> Result<(), String> {
         Ok(())
     } else {
         Err("Caller is not a known group canister".to_owned())
+    }
+}
+
+pub fn caller_is_video_call_operator() -> Result<(), String> {
+    if read_state(|state| state.is_caller_video_call_operator()) {
+        Ok(())
+    } else {
+        Err("Caller is not a video call operator".to_string())
     }
 }