+++ backend/canisters/user/impl/src/updates/c2c_notify_events.rs
@@ -2,6 +2,7 @@ use crate::guards::caller_is_local_user_index;
 use crate::{mutate_state, openchat_bot, RuntimeState};
 use canister_api_macros::update_msgpack;
 use canister_tracing_macros::trace;
+use chat_events::MessageContentInternal;
 use types::{MessageContentInitial, Timestamped};
 use user_canister::c2c_notify_events::{Response::*, *};
 use user_canister::mark_read::ChannelMessagesRead;
@@ -47,7 +48,12 @@ fn process_event(event: Event, state: &mut RuntimeState) {
         }
         Event::OpenChatBotMessage(content) => {
             let initial_content: MessageContentInitial = (*content).into();
-            openchat_bot::send_message(initial_content.try_into().unwrap(), false, state);
+            let now = state.env.now();
+            openchat_bot::send_message(
+                MessageContentInternal::from_initial(initial_content, now).unwrap(),
+                false,
+                state,
+            );
         }
         Event::UserJoinedGroup(ev) => {
             let now = state.env.now();