+++ backend/canisters/proposals_bot/CHANGELOG.md
@@ -9,6 +9,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 - Add `minimum_yes_proportion_of_total` to SNS proposals ([#5284](https://github.com/open-chat-labs/open-chat/pull/5284))
 
+### Fixed
+
+- Allow everyone to post in threads in new SNS proposals groups ([#5365](https://github.com/open-chat-labs/open-chat/pull/5365))
+- Update thread permissions in proposal groups ([#5373](https://github.com/open-chat-labs/open-chat/pull/5373))
+
 ## [[2.0.1027](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.1027-proposals_bot)] - 2024-01-25
 
 ### Changed